requirements.txt
python-dotenv
streamlit
transformers